fix: 공통 dto 및 component core 모듈로 이관
- management 모듈 공통 코드 이관
- build.gradle 의존 관계 정리
- management, collection 모듈 SecurityConfig 적용
