+++ auth/build.gradle
@@ -1,6 +1,9 @@
 dependencies {
     // core project
     implementation project(':core')
+    // spring
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    developmentOnly 'org.springframework.boot:spring-boot-devtools'
     // validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     // JPA
