+++ build.gradle
@@ -29,15 +29,6 @@ subprojects {
     dependencies {
         compileOnly 'org.projectlombok:lombok'
         annotationProcessor 'org.projectlombok:lombok'
-        // spring
-        implementation 'org.springframework.boot:spring-boot-starter-web'
-        developmentOnly 'org.springframework.boot:spring-boot-devtools'
-        // spring security
-        implementation 'org.springframework.boot:spring-boot-starter-security'
-        // kafka
-        implementation 'org.springframework.kafka:spring-kafka'
-        // JPA
-        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     }
 
     tasks.named('bootJar') {
