+++ collection/build.gradle
@@ -3,6 +3,11 @@ dependencies {
 	implementation project(':core')
 	// eureka
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
+	// spring
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	developmentOnly 'org.springframework.boot:spring-boot-devtools'
+	// spring security
+	implementation 'org.springframework.boot:spring-boot-starter-security'
 	// validation
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 	// H2
