+++ core/src/main/java/com/mongs/core/security/exception/ForbiddenHandler.java
@@ -4,18 +4,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mongs.core.error.ErrorResDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
-@Slf4j
-@Component
-@RequiredArgsConstructor
+
+@AllArgsConstructor
 public class ForbiddenHandler implements AccessDeniedHandler {
 
     private final ObjectMapper objectMapper;
