+++ core/src/main/java/com/mongs/core/security/exception/SecurityExceptionHandler.java
@@ -7,17 +7,13 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Component;
 import org.springframework.web.filter.GenericFilterBean;
 
 import java.io.IOException;
 
-@Slf4j
-@Component
-@RequiredArgsConstructor
+@AllArgsConstructor
 public class SecurityExceptionHandler extends GenericFilterBean {
 
     private final ObjectMapper objectMapper;
