+++ core/src/main/java/com/mongs/core/security/exception/UnAuthorizationHandler.java
@@ -4,18 +4,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mongs.core.error.ErrorResDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
-@Slf4j
-@Component
-@RequiredArgsConstructor
+@AllArgsConstructor
 public class UnAuthorizationHandler implements AuthenticationEntryPoint {
 
     private final ObjectMapper objectMapper;
