+++ core/src/main/java/com/mongs/core/security/filter/PassportFilter.java
@@ -1,0 +1,42 @@
+package com.mongs.core.security.filter;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mongs.core.passport.PassportVO;
+import com.mongs.core.security.principal.PassportDetail;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.AllArgsConstructor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.web.filter.GenericFilterBean;
+
+import java.io.IOException;
+
+@AllArgsConstructor
+public class PassportFilter extends GenericFilterBean {
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void doFilter(ServletRequest servletRequest, ServletResponse response, FilterChain chain) throws IOException, ServletException {
+        HttpServletRequest request = (HttpServletRequest) servletRequest;
+
+        String passportJson = request.getHeader("passport");
+
+        if (passportJson != null) {
+            PassportVO passportVO = objectMapper.readValue(passportJson, PassportVO.class);
+
+            User passport = new PassportDetail(passportVO);
+
+            UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken
+                    = new UsernamePasswordAuthenticationToken(passport, null, passport.getAuthorities());
+            SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
+        }
+
+        chain.doFilter(request, response);
+    }
+}
