--- core/src/main/java/com/mongs/core/security/passport/PassportFilter.java
@@ -1,43 +1,0 @@
-package com.mongs.core.security.passport;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mongs.core.passport.PassportVO;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.GenericFilterBean;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class PassportFilter extends GenericFilterBean {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        HttpServletRequest request = (HttpServletRequest) servletRequest;
-
-        String passportJson = request.getHeader("passport");
-
-        if (passportJson != null) {
-            PassportVO passportVO = objectMapper.readValue(passportJson, PassportVO.class);
-
-            User passport = new Passport(passportVO);
-
-            UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken
-                    = new UsernamePasswordAuthenticationToken(passport, null, passport.getAuthorities());
-            SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
-        }
-
-        chain.doFilter(request, response);
-    }
-}
