+++ gateway/build.gradle
@@ -5,7 +5,8 @@ dependencies {
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
 	// spring gateway
 	implementation 'org.springframework.cloud:spring-cloud-starter-gateway'
-	// gateway
+	developmentOnly 'org.springframework.boot:spring-boot-devtools'
+	// resilience4j
 	implementation 'org.springframework.cloud:spring-cloud-starter-circuitbreaker-resilience4j'
 	// jwt
 	implementation group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.5'
