--- gateway/src/test/java/com/mongs/gateway/GatewayApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mongs.gateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GatewayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
