+++ management/build.gradle
@@ -3,10 +3,19 @@ dependencies {
 	implementation project(':core')
 	// eureka
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
+	// spring
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	developmentOnly 'org.springframework.boot:spring-boot-devtools'
+	// spring security
+	implementation 'org.springframework.boot:spring-boot-starter-security'
 	// validation
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
+	// kafka
+	implementation 'org.springframework.kafka:spring-kafka'
 	// H2
 	runtimeOnly 'com.h2database:h2'
+	// JPA
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	// redis
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 	// test
