--- management/src/main/java/com/mongs/common/BasicCode.java
@@ -1,6 +1,0 @@
-package com.mongs.common;
-
-public interface BasicCode {
-    String getCode();
-    String getMessage();
-}
