+++ management/src/main/java/com/mongs/management/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.mongs.collection.config;
+package com.mongs.management.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mongs.core.security.exception.ForbiddenHandler;
@@ -38,8 +38,8 @@ public class SecurityConfig {
             .addFilterBefore(passportFilter, UsernamePasswordAuthenticationFilter.class)
             .addFilterBefore(securityExceptionHandler, PassportFilter.class)
             .authorizeHttpRequests(authorize -> authorize
-                    .requestMatchers("/collection/admin/**").hasAnyAuthority("ADMIN")
-                    .requestMatchers("/collection/**").hasAnyAuthority("NORMAL")
+                    .requestMatchers("/management/admin/**").hasAnyAuthority("ADMIN")
+                    .requestMatchers("/management/**").hasAnyAuthority("NORMAL")
                     .anyRequest().authenticated()
             )
             .exceptionHandling(configurer -> {
