--- management/src/main/java/com/mongs/management/management/repository/ManagementRepository.java
@@ -1,7 +1,0 @@
-package com.mongs.management.management.repository;
-
-import com.mongs.management.management.entity.Management;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ManagementRepository extends JpaRepository<Management, Long> {
-}
