--- management/src/main/java/com/mongs/management/management/service/dto/Evolution.java
@@ -1,18 +1,0 @@
-package com.mongs.management.management.service.dto;
-
-import lombok.*;
-
-/**
- * mongCode: String
- * stateCode: String
- * weight: Integer
- */
-
-
-@Builder
-public record Evolution (
-    String mongCode,
-    String stateCode,
-    double weight
-){
-}
