--- management/src/main/java/com/mongs/management/management/service/dto/FeedCode.java
@@ -1,12 +1,0 @@
-package com.mongs.management.management.service.dto;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-public record FeedCode (
-    String feedCode
-){
-}
