--- management/src/main/java/com/mongs/management/management/service/dto/Graduation.java
@@ -1,10 +1,0 @@
-package com.mongs.management.management.service.dto;
-
-import lombok.*;
-
-@Builder
-public record Graduation(
-        String mongCode
-)
-{
-}
