--- management/src/main/java/com/mongs/management/management/service/dto/InitMong.java
@@ -1,17 +1,0 @@
-package com.mongs.management.management.service.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * name: string
- * sleepStart: string (HH:mm)
- * sleepEnd: string (HH:mm)
- */
-public record InitMong (
-    String name,
-    String sleepStart,
-    String sleepEnd
-){
-}
