--- management/src/main/java/com/mongs/management/management/service/dto/TrainingCount.java
@@ -1,5 +1,0 @@
-package com.mongs.management.management.service.dto;
-
-public class TrainingCount {
-    private int trainingCount;
-}
