+++ management/src/main/java/com/mongs/management/repository/ManagementRepository.java
@@ -1,0 +1,7 @@
+package com.mongs.management.repository;
+
+import com.mongs.management.entity.Management;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ManagementRepository extends JpaRepository<Management, Long> {
+}
