+++ management/src/main/java/com/mongs/management/service/dto/CreateMong.java
@@ -1,0 +1,25 @@
+package com.mongs.management.service.dto;
+
+
+import com.mongs.management.exception.ManagementErrorCode;
+import com.mongs.management.entity.Management;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Builder
+public record CreateMong (
+    String name,
+    String mongCode,
+    double weight,
+    LocalDateTime borntime
+){
+    public static CreateMong of(Management management) {
+        return CreateMong.builder()
+                .name(management.getName())
+                .mongCode(ManagementErrorCode.SUCCESS.getCode())
+                .weight(management.getWeight())
+                .build();
+    }
+
+}
