+++ management/src/main/java/com/mongs/management/service/dto/Evolution.java
@@ -1,0 +1,18 @@
+package com.mongs.management.service.dto;
+
+import lombok.*;
+
+/**
+ * mongCode: String
+ * stateCode: String
+ * weight: Integer
+ */
+
+
+@Builder
+public record Evolution (
+    String mongCode,
+    String stateCode,
+    double weight
+){
+}
