+++ management/src/main/java/com/mongs/management/service/dto/FeedCode.java
@@ -1,0 +1,9 @@
+package com.mongs.management.service.dto;
+
+import lombok.Builder;
+
+@Builder
+public record FeedCode (
+    String feedCode
+){
+}
