+++ management/src/main/java/com/mongs/management/service/dto/Graduation.java
@@ -1,0 +1,10 @@
+package com.mongs.management.service.dto;
+
+import lombok.*;
+
+@Builder
+public record Graduation(
+        String mongCode
+)
+{
+}
