+++ management/src/main/java/com/mongs/management/service/dto/TrainingCount.java
@@ -1,0 +1,5 @@
+package com.mongs.management.service.dto;
+
+public class TrainingCount {
+    private int trainingCount;
+}
